source/twitter_data_mining/requirements.txt
cffi==1.9.1
cryptography==1.7.1
enum34==1.1.6
idna==2.2
ipaddress==1.0.17
kafka-python==1.3.2
oauthlib==2.0.1
pyasn1==0.1.9
pycparser==2.17
pyOpenSSL==16.2.0
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
tweepy==3.5.0
